# openapi.yaml
openapi: 3.0.0
info:
  title: Chronos API
  version: 1.0.0
  description: API for Chronos Medlemsregister
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/members:
    get:
      operationId: getMembers
      summary: Get members
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-Org-ID
          schema:
            type: string
          description: Optional organization filter
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
  /api/organizations:
    get:
      operationId: getOrganizations
      summary: Get unique organization IDs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of organization IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
  /api/reports/as-of:
    get:
      operationId: getMembersAsOf
      summary: Get members as of a specific date
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
        - in: header
          name: X-Org-ID
          schema:
            type: string
      responses:
        '200':
          description: List of members as of date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
...
  /api/forms:
    get:
      operationId: getForms
      summary: Get available forms
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: org_id
          schema:
            type: string
      responses:
        '200':
          description: List of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Form'
...
  /api/commands/create-form:
    post:
      operationId: createForm
      summary: Create a new form
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormRequest'
      responses:
        '201':
          description: Form created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Form:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        title:
          type: string
        schema:
          type: object
          additionalProperties: true
...
    CreateFormRequest:
      type: object
      required:
        - org_id
        - title
        - schema
      properties:
        org_id:
          type: string
        title:
          type: string
        schema:
          type: object
          additionalProperties: true
...
    Member:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
        metadata:
          type: object
          additionalProperties: true
        balance:
          type: integer
        fee_formula:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...